Move gif selector searchbox
